function/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
azure-storage-blob
langchain
langchain-pinecone
langchain-openai
# langchain-community
# pinecone-notebooks
# sentence-transformers
unstructured-client
unstructured
pinecone-client
openai
psutil
python-dotenv